relocated generatePinTimestamp to pin service
