+++ admin/services/check-start.service/check-start.service.js
@@ -17,7 +17,7 @@ const checkFormService = require('../check-form.service')
 const configService = require('../config.service')
 const dateService = require('../date.service')
 const pinGenerationDataService = require('../data-access/pin-generation.data.service')
-const pinValidityGeneratorService = require('../pin-validity-generator.service')
+const pinService = require('../pin.service')
 const prepareCheckService = require('../prepare-check.service')
 const queueNameService = require('../queue-name-service')
 const sasTokenService = require('../sas-token.service')
@@ -210,7 +210,7 @@ checkStartService.initialisePupilCheck = async function (
     checkWindow_id: checkWindow.id,
     isLiveCheck: isLiveCheck
   }
-  checkData.pinExpiresAt = pinValidityGeneratorService.generatePinTimestamp(config.OverridePinExpiry, endOfDay(), fourPmToday(), schoolTimezone)
+  checkData.pinExpiresAt = pinService.generatePinTimestamp(config.OverridePinExpiry, endOfDay(), fourPmToday(), schoolTimezone)
   checkData.school_id = schoolId
 
   // checkCode will be created by the database on insert
