+++ admin/spec/back-end/service/prepare-check.service.spec.js
@@ -2,7 +2,7 @@
 
 /* global describe it expect beforeEach spyOn fail */
 const sut = require('../../../services/prepare-check.service')
-const pinValidityGeneratorService = require('../../../services/pin-validity-generator.service')
+const pinService = require('../../../services/pin.service')
 const redisService = require('../../../services/data-access/redis-cache.service')
 const moment = require('moment')
 
@@ -10,7 +10,7 @@ let check
 
 describe('prepare-check.service', () => {
   beforeEach(() => {
-    spyOn(pinValidityGeneratorService, 'generatePinTimestamp')
+    spyOn(pinService, 'generatePinTimestamp')
     check = {
       checkCode: 'check-code',
       schoolPin: 'school-pin',
